Split standardrb to its own CI step (#102)
Currently, CI runs standardrb and then tests, so if standardb fails you
don't get any feedback from the test.

It also runs standardb on all 3 versions of ruby. In practice, you
really only need to run it on one version.

This splits `linters` out from `tests`. I tried to mirror the naming,
but I'm open to explicitly calling it standardrb.

Lastly, I'm trying to turn on `--format github` which HOPEFULLY will
start annotating the PR with the failure. We shouldn't have to rely on
a maintainer to point out that standardrb --fix is needed :sweat_smile:

Oh, and since we are here, also turn on color output since GitHub
Actions supports it.

Co-authored-by: Andrei Bondarev <andrei@sourcelabs.io>